chore(ci): satisfy cloud workspace lint
