+++ src/gateway/worker-environments/workspace-reconcile.ts
@@ -1,13 +1,10 @@
 import { createHash, randomBytes } from "node:crypto";
-import { createReadStream } from "node:fs";
 import fs from "node:fs/promises";
 import os from "node:os";
 import path from "node:path";
-import { FsSafeError, root as openFsSafeRoot, type Root } from "../../infra/fs-safe.js";
 import { runCommandBuffered } from "../../process/exec.js";
 import { runCommandWithTimeout } from "../../process/exec.js";
 import {
-  gitFileMode,
   MAX_RECONCILIATION_ENTRIES,
   MAX_RECONCILIATION_FILE_BYTES,
   MAX_RECONCILIATION_TOTAL_BYTES,
@@ -17,6 +14,20 @@ import {
   type WorkerWorkspaceReconciliationJournalAdapter,
 } from "./workspace-manifest.js";
 import { isDerivedWorkspacePath } from "./workspace-path-exclusions.js";
+import {
+  prepareNonDirectoryTargets,
+  reconciliationDirectories,
+  reconciliationEntries,
+} from "./workspace-reconcile-derived-paths.js";
+import {
+  absoluteEntryMatches,
+  clearTemporaryWorkspace,
+  directoryContainsOnlyDerivedWorkspaceEntries,
+  directoryContainsOnlyJournalPaths,
+  entryMatches,
+  localPath,
+  readWorkspaceTreeFile,
+} from "./workspace-reconcile-fs.js";
 export {
   MAX_RECONCILIATION_ENTRIES,
   MAX_RECONCILIATION_FILE_BYTES,
@@ -31,138 +42,6 @@ export {
 const PATCH_TIMEOUT_MS = 10 * 60_000;
 
 class ConcurrentWorkspacePathError extends Error {}
-
-function reconciliationEntries(
-  entries: readonly WorkerWorkspaceManifestEntry[],
-): WorkerWorkspaceManifestEntry[] {
-  return entries.filter((entry) => !isDerivedWorkspacePath(entry.path));
-}
-
-function reconciliationDirectories(directories: readonly string[] | undefined): string[] {
-  return (directories ?? []).filter((directory) => !isDerivedWorkspacePath(directory));
-}
-
-function localPath(root: string, relative: string): string {
-  return path.join(root, ...relative.split("/"));
-}
-
-async function sha256File(filePath: string): Promise<string> {
-  const hash = createHash("sha256");
-  for await (const chunk of createReadStream(filePath)) {
-    hash.update(chunk);
-  }
-  return hash.digest("hex");
-}
-
-async function absoluteEntryMatches(
-  absolute: string,
-  entry: WorkerWorkspaceManifestEntry,
-): Promise<boolean> {
-  const stats = await fs.lstat(absolute).catch(() => undefined);
-  if (!stats) {
-    return false;
-  }
-  if (entry.type === "symlink") {
-    return stats.isSymbolicLink() && (await fs.readlink(absolute)) === entry.target;
-  }
-  return (
-    stats.isFile() &&
-    !stats.isSymbolicLink() &&
-    gitFileMode(stats.mode & 0o777) === entry.mode &&
-    stats.size === entry.size &&
-    (await sha256File(absolute)) === entry.sha256
-  );
-}
-
-async function entryMatches(root: string, entry: WorkerWorkspaceManifestEntry): Promise<boolean> {
-  return await absoluteEntryMatches(localPath(root, entry.path), entry);
-}
-
-async function removeDerivedWorkspaceDescendants(
-  root: Root,
-  relativeDirectory: string,
-): Promise<void> {
-  for (const entry of await root.list(relativeDirectory, { withFileTypes: true })) {
-    const child = relativeDirectory ? `${relativeDirectory}/${entry.name}` : entry.name;
-    if (isDerivedWorkspacePath(child)) {
-      await removeDerivedWorkspaceEntry(root, child, entry.isDirectory);
-      continue;
-    }
-    if (entry.isDirectory) {
-      await removeDerivedWorkspaceDescendants(root, child);
-    }
-  }
-}
-
-async function removeDerivedWorkspaceEntry(
-  root: Root,
-  relativePath: string,
-  isDirectory: boolean,
-): Promise<void> {
-  if (isDirectory) {
-    let entries;
-    try {
-      entries = await root.list(relativePath, { withFileTypes: true });
-    } catch (error) {
-      if (!(error instanceof FsSafeError) || !["not-found", "path-alias"].includes(error.code)) {
-        throw error;
-      }
-      entries = undefined;
-    }
-    for (const entry of entries ?? []) {
-      await removeDerivedWorkspaceEntry(root, `${relativePath}/${entry.name}`, entry.isDirectory);
-    }
-  }
-  await root.remove(relativePath).catch((error: unknown) => {
-    if (!(error instanceof FsSafeError) || error.code !== "not-found") {
-      throw error;
-    }
-  });
-}
-
-async function hasWorkspaceSymlinkAncestor(root: string, relativePath: string): Promise<boolean> {
-  const segments = relativePath.split("/");
-  for (let index = 1; index < segments.length; index += 1) {
-    const stats = await fs
-      .lstat(localPath(root, segments.slice(0, index).join("/")))
-      .catch(() => undefined);
-    if (stats?.isSymbolicLink()) {
-      return true;
-    }
-  }
-  return false;
-}
-
-async function prepareNonDirectoryTargets(
-  root: string,
-  entries: readonly WorkerWorkspaceManifestEntry[],
-): Promise<void> {
-  const workspaceRoot = await openFsSafeRoot(root);
-  for (const entry of reconciliationEntries(entries)) {
-    if (await hasWorkspaceSymlinkAncestor(root, entry.path)) {
-      continue;
-    }
-    let stats;
-    try {
-      stats = await workspaceRoot.stat(entry.path);
-    } catch (error) {
-      if (error instanceof FsSafeError && ["not-found", "path-alias"].includes(error.code)) {
-        continue;
-      }
-      throw error;
-    }
-    if (stats.isDirectory) {
-      // Excluded descendants cannot fence a namespace replacement, but Git
-      // cannot replace their still-nonempty parent until they are removed.
-      // fs-safe binds traversal/removal to root-relative directory handles, so
-      // a symlink swap cannot redirect cleanup outside this workspace.
-      await removeDerivedWorkspaceDescendants(workspaceRoot, entry.path);
-      if ((await workspaceRoot.list(entry.path)).length === 0) {
-        await workspaceRoot.remove(entry.path);
-      }
-    }
-  }
-}
 
 export async function assertWorkspaceMatchesManifest(params: {
   root: string;
@@ -446,57 +325,6 @@ async function writeRawWorkspaceTree(params: {
   return await requireGit(params.repositoryRoot, ["rev-parse", `${ref}^{tree}`]);
 }
 
-async function readWorkspaceTreeFile(params: {
-  repositoryRoot: string;
-  tree: string;
-  entry: Extract<WorkerWorkspaceManifestEntry, { type: "file" }>;
-}): Promise<Uint8Array> {
-  const listed = await runCommandBuffered(
-    [
-      "git",
-      "--literal-pathspecs",
-      "-C",
-      params.repositoryRoot,
-      "ls-tree",
-      "-z",
-      "--full-tree",
-      params.tree,
-      "--",
-      params.entry.path,
-    ],
-    {
-      timeoutMs: PATCH_TIMEOUT_MS,
-      maxOutputBytes: 1024 * 1024,
-    },
-  );
-  if (listed.termination !== "exit" || listed.code !== 0) {
-    throw new Error(listed.stderr.toString("utf8").trim() || "git ls-tree failed");
-  }
-  const record = listed.stdout;
-  const terminator = record.indexOf(0);
-  const separator = record.indexOf(9);
-  if (terminator !== record.byteLength - 1 || separator < 0 || separator > terminator) {
-    throw new Error(`Cloud workspace recovery snapshot is missing: ${params.entry.path}`);
-  }
-  const metadata = record.subarray(0, separator).toString("utf8");
-  const match = /^100(?:644|755) blob ([a-f0-9]{40})$/u.exec(metadata);
-  const listedPath = record.subarray(separator + 1, terminator);
-  if (!match || !listedPath.equals(Buffer.from(params.entry.path))) {
-    throw new Error(`Cloud workspace recovery snapshot is invalid: ${params.entry.path}`);
-  }
-  const blob = await runCommandBuffered(
-    ["git", "-C", params.repositoryRoot, "cat-file", "blob", match[1]!],
-    {
-      timeoutMs: PATCH_TIMEOUT_MS,
-      maxOutputBytes: MAX_RECONCILIATION_FILE_BYTES + 1,
-    },
-  );
-  if (blob.termination !== "exit" || blob.code !== 0) {
-    throw new Error(blob.stderr.toString("utf8").trim() || "git cat-file failed");
-  }
-  return blob.stdout;
-}
-
 async function createWorkspacePatch(params: {
   root: string;
   stagingRoot: string;
@@ -629,48 +457,6 @@ function validateJournalSnapshot(journal: WorkerWorkspaceReconciliationJournal):
     createHash("sha256").update(journal.basePack).digest("hex") !== journal.basePackSha256
   ) {
     throw new Error("Cloud workspace reconciliation recovery snapshot is invalid");
-  }
-}
-
-async function directoryContainsOnlyJournalPaths(
-  root: string,
-  directory: string,
-  paths: ReadonlySet<string>,
-  directories: ReadonlySet<string>,
-): Promise<boolean> {
-  for (const name of await fs.readdir(localPath(root, directory))) {
-    const child = `${directory}/${name}`;
-    if (isDerivedWorkspacePath(child)) {
-      continue;
-    }
-    const stats = await fs.lstat(localPath(root, child));
-    if (stats.isDirectory() && !stats.isSymbolicLink()) {
-      if (!directories.has(child)) {
-        return false;
-      }
-      if (!(await directoryContainsOnlyJournalPaths(root, child, paths, directories))) {
-        return false;
-      }
-    } else if (!paths.has(child)) {
-      return false;
-    }
-  }
-  return true;
-}
-
-async function directoryContainsOnlyDerivedWorkspaceEntries(
-  root: string,
-  directory: string,
-): Promise<boolean> {
-  const names = await fs.readdir(localPath(root, directory));
-  return names.length > 0 && names.every((name) => isDerivedWorkspacePath(`${directory}/${name}`));
-}
-
-async function clearTemporaryWorkspace(repositoryRoot: string): Promise<void> {
-  for (const name of await fs.readdir(repositoryRoot)) {
-    if (name !== ".git") {
-      await fs.rm(path.join(repositoryRoot, name), { recursive: true, force: true });
-    }
   }
 }
