+++ src/gateway/worker-environments/workspace-sync-scripts.ts
@@ -4,38 +4,7 @@ import {
   DERIVED_WORKSPACE_FILE_SUFFIXES,
   isDerivedWorkspacePath,
 } from "./workspace-path-exclusions.js";
-
-export const REMOTE_WORKSPACE_SETUP_SCRIPT = String.raw`set -eu
-relative=$1
-root=$HOME/.openclaw-worker
-
-ensure_private_directory() {
-  directory=$1
-  if [ -e "$directory" ] || [ -L "$directory" ]; then
-    if [ ! -d "$directory" ] || [ -L "$directory" ]; then
-      printf '%s\n' 'unsafe worker workspace directory' >&2
-      exit 2
-    fi
-  else
-    mkdir "$directory"
-  fi
-  chmod 700 "$directory"
-}
-
-ensure_private_directory "$root"
-current=$root
-old_ifs=$IFS
-IFS=/
-set -- $relative
-IFS=$old_ifs
-for segment in "$@"; do
-  current=$current/$segment
-  ensure_private_directory "$current"
-done
-cd "$current"
-find . -mindepth 1 -maxdepth 1 -exec rm -rf -- {} +
-pwd -P
-`;
+export { REMOTE_WORKSPACE_SETUP_SCRIPT } from "./workspace-sync-setup-script.js";
 
 export const REMOTE_WORKSPACE_QUIESCE_JS = String.raw`const childProcess = require("node:child_process");
 const crypto = require("node:crypto");
