+++ src/gateway/worker-environments/workspace-sync-setup-script.ts
@@ -1,0 +1,31 @@
+export const REMOTE_WORKSPACE_SETUP_SCRIPT = String.raw`set -eu
+relative=$1
+root=$HOME/.openclaw-worker
+
+ensure_private_directory() {
+  directory=$1
+  if [ -e "$directory" ] || [ -L "$directory" ]; then
+    if [ ! -d "$directory" ] || [ -L "$directory" ]; then
+      printf '%s\n' 'unsafe worker workspace directory' >&2
+      exit 2
+    fi
+  else
+    mkdir "$directory"
+  fi
+  chmod 700 "$directory"
+}
+
+ensure_private_directory "$root"
+current=$root
+old_ifs=$IFS
+IFS=/
+set -- $relative
+IFS=$old_ifs
+for segment in "$@"; do
+  current=$current/$segment
+  ensure_private_directory "$current"
+done
+cd "$current"
+find . -mindepth 1 -maxdepth 1 -exec rm -rf -- {} +
+pwd -P
+`;
